+++ app/api/auth/[...nextauth]/route.ts
@@ -1,54 +1,4 @@
-import NextAuth, { NextAuthOptions } from "next-auth";
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
-import prisma from "@/lib/prisma";
-import GoogleProvider from "next-auth/providers/google";
-import { GetUserByID } from "@/app/api/auth/lib/repository";
-import { Role } from "@/types/next-auth";
+import NextAuth from "next-auth";
+import { authOptions } from "@/app/server/auth";
 
-export const authOptions: NextAuthOptions = {
-  adapter: PrismaAdapter(prisma),
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID as string,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
-    }),
-  ],
-  callbacks: {
-    async signIn({ user }) {
-      GetUserByID(user.id)
-        .then((result) => {
-          if (result?.role) {
-            switch (result.role) {
-              case "admin":
-                user.role = Role.admin;
-                break;
-              case "user":
-                user.role = Role.user;
-                break;
-            }
-            return true;
-          }
-        })
-        .catch((error) => {
-          console.error(error);
-        });
-      return false;
-    },
-    async jwt({ token, user }) {
-      if (user) {
-        token.role = user.role;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      if (session.user && token) {
-        session.user.role = token.role;
-      }
-      return session;
-    },
-  },
-};
-
-const handler = NextAuth(authOptions);
-
-export { handler as GET, handler as POST };
+export default NextAuth(authOptions);
