+++ app/auth/signin/page.tsx
@@ -1,0 +1,19 @@
+import type { InferGetServerSidePropsType } from "next";
+import { signIn } from "next-auth/react";
+import { getServerSideProps } from "@/app/server/auth";
+
+export default function SignIn({
+  providers,
+}: InferGetServerSidePropsType<typeof getServerSideProps>) {
+  return (
+    <>
+      {Object.values(providers).map((provider) => (
+        <div key={provider.name}>
+          <button onClick={() => signIn(provider.id)}>
+            Sign in with {provider.name}
+          </button>
+        </div>
+      ))}
+    </>
+  );
+}
