--- types/next-auth.d.ts
@@ -1,19 +1,0 @@
-import NextAuth, { DefaultUser } from "next-auth";
-
-export enum Role {
-  user = "user",
-  admin = "admin",
-}
-
-interface IUser extends DefaultUser {
-  role?: Role;
-}
-declare module "next-auth" {
-  interface User extends IUser {}
-  interface Session {
-    user?: User;
-  }
-}
-declare module "next-auth/jwt" {
-  interface JWT extends IUser {}
-}
